pyproject.toml
[tool.black]
line-length = 120
target-version = ["py310"]
include = ".*py$"

[tool.isort]
profile = "black"
force_single_line = false

[tool.poetry]
name = "snedbot"
version = "0.1.0" # I do not actually update this, lol
description = "Your friendly neighbourhood multi-purpose Discord bot."
authors = ["HyperGH <46067571+HyperGH@users.noreply.github.com>"]
license = "GNU GPL-v3"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
dateparser = "^1.1.4"
psutil = "^5.9.4"
Pillow = "^9.3.0"
asyncpg = "^0.27.0"
Levenshtein = "^0.20.8"
uvloop = {version = "==0.17.0", platform="linux"}
aiodns = "~=3.0"
cchardet = "~=2.1"
Brotli = "~=1.0"
ciso8601 = "~=2.2"
kosu = {git = "https://github.com/HyperGH/kosu.git", branch="main"}
hikari-lightbulb = "==2.3.0"
hikari-miru = "==2.0.1"

[tool.poetry.dev-dependencies]
black = "^22.10"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
